fix: Moved assets and fixed tests
