+++ src/assets/vocabulary.json
