+++ src/targets/services/helpers/index.js
@@ -1,5 +1,5 @@
-import vocabulary from './vocabulary_tiny.json'
-import classes from './classes.json'
+import vocabulary from '../../../assets/vocabulary_tiny.json'
+import classes from '../../../assets/classes.json'
 
 export { Model } from './model'
 export { vocabulary, classes }
